python_agent/pyproject.toml
[project]
name = "python-agent"
version = "0.1.0"
description = "Financial AI Agent with gRPC and LangGraph"
authors = [
    { name = "ZXQ", email = "you@example.com" }
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "langchain-openai>=1.2.0",
    "langgraph>=0.1.0",
    "python-dotenv>=1.0.0",
    "langchain-community (>=0.4.1,<0.5.0)",
    "tavily-python (>=0.7.23,<0.8.0)",
    "streamlit (>=1.56.0,<2.0.0)",
    "langgraph-checkpoint-sqlite (>=3.0.3,<4.0.0)",
    "langgraph-checkpoint (>=4.0.2,<5.0.0)",
    "langchain-tavily (>=0.2.18,<0.3.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "chromadb (>=1.5.8,<2.0.0)",
    "sentence-transformers (>=5.4.1,<6.0.0)",
    "yfinance (>=1.3.0,<2.0.0)",
    "ccxt (>=4.5.50,<5.0.0)"
]

[tool.poetry]
# 核心关键：通常不需要打包发布，必须设为 false
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"